vscode: add configuration
This adds a Daedalus configuration section for the VSCode extension.  For now
there is just one option "daedalus.language-server-path", to set the path of
the executable that should be run as a language server.

Note that extension.ts uses Windows-style line endings (CRLF), which I think I
am respecting in my changes, but in this commit, the new configuration.ts uses
Unix-style line endings (LF).  Should CRLF be preferred?

I also removed comments that seem like they were copy-pasted from some example
code and are not accurate to this project.