Change how we resolve tycons in signatures.
If we are not known types, we just look for the corresponding
rules and we use an instantiated version of the result type.

These seems more stable, and it has the nice benefit of allowing
us to refer to the results of rules, even of they don't introduce
a new type.