Codegen for bitdata union in HS, and change typing of {}
Previously {} always had type {}.  With bitdata it is possible
to define empty structs that are not `{}`, for example:

bitdata A where
   1 : uint 8

More commonly this happens with the (anonymous) fields of union
bitdata.   To work around this we treat the value `{}` as an empty
struct (i.e., it can work for any struct type that has no fields).

If we end up generating fresh types though, we default such empty
struct to the unit type {}