Checkpoing: Track which names might be visible to external users
When compiling, such names are printed more or less "as is".  Names that
are not visible otoh would typically get some sort of unique identifier
appended to them to avoid clashes.

XXX: The Haskell compiler needs to repsect this
XXX: When printing FNames in user errors we should no append the unique
     identifier.