formats/http: parse Content-Length and Transfer-Encoding headers
This change extends the HTTP parser with support for parsing the
Content-Length and Transfer-Encoding headers according to the HTTP
specification. Other headers are represented using a catch-all 'Header'
struct. We're parsing the two specific headers mentioned here to prepare
the parser so that later changes can parse the request body according to
whether the transfer encoding is 'chunked'.

This change makes the simplifying assumption that Content-Length values
are representable in at most 64 bits of storage. While the specification
does not place an upper limit on the size of Content-Length values, we
impose a limit out of practicality.