formats/http: update chunked encoding handling to honor encoding position
This change updates the HTTP message parser to respect chunked encoding
only when it is the last encoding in the encoding list, as specified. In
addition, this change updates the parser to handle the case when the
body length is indeterminate, i.e, when the body should be obtained by
reading all remaining available bytes from the input stream, as
specified.