formats/http: migrate the examples and example runner script to a test suite
The test suite is run by running run-tests.sh, and each test case lives
in tests/inputs/. The test suite runner works by running each input file
using the HTTP spec and then comparing the output to an expected output
file in tests/outputs/, named with the same name as its corresponding
input file. A diff of the results is shown if there are differences.