formats/http: add the beginning of an HTTP 2 parser along with support for testing it
This change adds the beginning of an HTTP 2 frame parser. This change is
mostly focused on extending the HTTP parser test suite to support HTTP
2, which requires that we provide some relatively humane way to write
test cases. The approach taken here was to allow users to write binary
test case inputs in "xxd" format. The test suite then converts those
into binary on the fly when running the tests. See run-tests.sh and
README.rst for more details.