formats/http: simplify settings frame parsing
This change uses a more idiomatic approach to parsing settings frames.
Rather than explicitly checking for a body of the expected size, we
parse the body in a constrained "Chunk" and expect that all input in the
declared frame length will be consumed. This change also modifies the
parser so that rather than failing on unknown settings identifiers, we
return them to the application in a new SETTINGS_UNKNOWN identifier
constructor. This means that it is the application's responsibility to
ignore such settings, although that should be straightforward since
they'll be tagged in this way.