Don't allow double quotes on the content length.
The spec on this is not very clear, because in one place it says that
quotes are OK everywhere, but in the definition for content length it
only has digits.   There does not seem to be a good reason to have the
length in quotes, unlike other places where the values may contain white
space.