Adds PHP Archive (PHAR) format and polyglots
The PHP Archive format is of special interest because it admits a prefix
cavity, forming the base of stack polyglot skeletons to produce polyglots with
JPG, GIF, and many other common file formats.

This PR defines a PHAR Daedalus format and adds examples of most Mitra
stack-based polyglots.

Also includes utilities for generating valid PHAR polyglots, which requires
adding a PHAR signature as a postprocessing step.