BUGFIX: Let terms in CFG creation
Fixes what I suspect is a bug in how `Let` terms are translated to CFG nodes,
where the bound name is inserted into the wrong node(s).

Seems like a benign bug though -- I think `Let` terms are translated to `Do`
terms somewhere upstream, because I wasn't able to reproduce the bug in
practice, and I confirmed that this code is not executed when processing a test
file containing `let` statements.